checking in more progess, 0DRO somewhat working
