+++ watertap/core/membrane_channel_base.py
@@ -14,9 +14,9 @@
 from copy import deepcopy
 from enum import Enum, auto
 
-from pyomo.common.collections import ComponentSet
 from pyomo.common.config import Bool, ConfigDict, ConfigValue, In
 from pyomo.environ import (
+    Expression,
     Param,
     NegativeReals,
     NonNegativeReals,
@@ -37,6 +37,7 @@ from idaes.core.solvers import get_solver
 from idaes.core.util import scaling as iscale
 from idaes.core.util.config import is_physical_parameter_block
 from idaes.core.util.exceptions import ConfigurationError, InitializationError
+from idaes.core.util.misc import add_object_reference
 
 import idaes.logger as idaeslog
 
@@ -267,11 +268,20 @@ CONFIG_Template.declare(
 
 class MembraneChannelMixin:
 
-    def add_mass_transfer(self):
-        raise NotImplementedError()
-
     def _add_pressure_change(self, pressure_change_type=PressureChangeType.calculated):
         raise NotImplementedError()
+
+    def _add_reference(self, name, var):
+        if not isinstance(var, (Var, Param, Expression)):
+            raise ConfigurationError(
+                f"{self.name} var must be a Pyomo Var, Param or "
+                "Expression."
+            )
+        elif var.is_indexed():
+            raise ConfigurationError(
+                f"{self.name} var must be a scalar (unindexed) " "component."
+            )
+        add_object_reference(self, name, var)
 
     def _add_pressure_change_equation(self):
         @self.Constraint(
@@ -280,57 +290,6 @@ class MembraneChannelMixin:
         def eq_pressure_change(b, t):
             return b.pressure_change_total[t] == sum(
                 b.dP_dx[t, x] * b.length / b.nfe for x in b.difference_elements
-            )
-
-    def _add_recovery_rejection(self,**kwrags):
-
-        solvent_set = self.config.property_package.solvent_set
-        solute_set = self.config.property_package.solute_set
-
-        self.recovery_mass_phase_comp = Var(
-            self.flowsheet().config.time,
-            self.config.property_package.phase_list,
-            self.config.property_package.component_list,
-            initialize=lambda b, t, p, j: 0.4037 if j in solvent_set else 0.0033,
-            bounds=lambda b, t, p, j: (1e-2, 1 - 1e-6)
-            if j in solvent_set
-            else (1e-5, 1 - 1e-6),
-            units=pyunits.dimensionless,
-            doc="Mass-based component recovery",
-        )
-
-        self.rejection_phase_comp = Var(
-            self.flowsheet().config.time,
-            self.config.property_package.phase_list,
-            solute_set,
-            initialize=0.9,
-            bounds=(1e-2, 1 - 1e-6),
-            units=pyunits.dimensionless,
-            doc="Observed solute rejection",
-        )
-
-        # ==========================================================================
-        # Mass-based Component Recovery rate
-        @self.Constraint(
-            self.flowsheet().config.time, self.config.property_package.component_list
-        )
-        def eq_recovery_mass_phase_comp(b, t, j):
-            return (
-                b.recovery_mass_phase_comp[t, "Liq", j]
-                * b.properties[t, b.first_element].flow_mass_phase_comp[
-                    "Liq", j
-                ]
-                == b.mixed_permeate[t].flow_mass_phase_comp["Liq", j]
-            )
-
-        # rejection
-        @self.Constraint(self.flowsheet().config.time, solute_set)
-        def eq_rejection_phase_comp(b, t, j):
-            return b.rejection_phase_comp[t, "Liq", j] == 1 - (
-                b.mixed_permeate[t].conc_mass_phase_comp["Liq", j]
-                / b.properties[t, self.first_element].conc_mass_phase_comp[
-                    "Liq", j
-                ]
             )
 
     def add_total_pressure_balances(
@@ -343,7 +302,13 @@ class MembraneChannelMixin:
             has_pressure_change=has_pressure_change,
             custom_term=custom_term)
 
-        self._add_membrane_pressure_balances()
+        @self.Constraint(
+            self.flowsheet().config.time,
+            self.difference_elements,
+            doc="Pressure at interface",
+        )
+        def eq_equal_pressure_interface(b, t, x):
+            return b.properties_interface[t, x].pressure == b.properties[t, x].pressure
 
         if has_pressure_change:
             self._add_pressure_change(pressure_change_type=pressure_change_type)
@@ -352,19 +317,6 @@ class MembraneChannelMixin:
             self._add_calculated_pressure_change()
 
     def add_isothermal_conditions(self):
-
-        # # ==========================================================================
-        # Feed and permeate-side isothermal conditions
-        @self.Constraint(
-            self.flowsheet().config.time,
-            self.length_domain,
-            doc="Isothermal assumption for permeate",
-        )
-        def eq_permeate_isothermal(b, t, x):
-            return (
-                b.properties[t, x].temperature
-                == b.permeate_side[t, x].temperature
-            )
 
         # ==========================================================================
         # Bulk and interface connections on the feed-side
@@ -380,18 +332,6 @@ class MembraneChannelMixin:
                 == b.properties_interface[t, x].temperature
             )
 
-        # ==========================================================================
-        # isothermal conditions at permeate outlet
-        @self.Constraint(
-            self.flowsheet().config.time, doc="Isothermal assumption for permeate out"
-        )
-        def eq_permeate_outlet_isothermal(b, t):
-            return (
-                b.properties[t, b.length_domain.last()].temperature
-                == b.mixed_permeate[t].temperature
-            )
-
-
     def add_volumetric_flowrate_balance(self):
         # VOLUMETRIC FLOWRATE
         @self.Constraint(
@@ -404,103 +344,6 @@ class MembraneChannelMixin:
                 b.properties_interface[t, x].flow_vol_phase["Liq"]
                 == b.properties[t, x].flow_vol_phase["Liq"]
             )
-
-
-    def add_flux_balance(self):
-
-        solvent_set = self.config.property_package.solvent_set
-        solute_set = self.config.property_package.solute_set
-
-        units_meta = self.config.property_package.get_metadata().get_derived_units
-
-        self.A_comp = Var(
-            self.flowsheet().config.time,
-            solvent_set,
-            initialize=1e-12,
-            bounds=(1e-18, 1e-6),
-            domain=NonNegativeReals,
-            units=units_meta("length")
-            * units_meta("pressure") ** -1
-            * units_meta("time") ** -1,
-            doc="Solvent permeability coeff.",
-        )
-
-        self.B_comp = Var(
-            self.flowsheet().config.time,
-            solute_set,
-            initialize=1e-8,
-            bounds=(1e-11, 1e-5),
-            domain=NonNegativeReals,
-            units=units_meta("length") * units_meta("time") ** -1,
-            doc="Solute permeability coeff.",
-        )
-
-        # TODO: add water density to NaCl prop model and remove here (or use IDAES version)
-        self.dens_solvent = Param(
-            initialize=1000,
-            units=units_meta("mass") * units_meta("length") ** -3,
-            doc="Pure water density",
-        )
-
-        self.flux_mass_phase_comp = Var(
-            self.flowsheet().config.time,
-            self.length_domain,
-            self.config.property_package.phase_list,
-            self.config.property_package.component_list,
-            initialize=lambda b, t, x, p, j: 5e-4 if j in solvent_set else 1e-6,
-            bounds=lambda b, t, x, p, j: (1e-4, 3e-2)
-            if j in solvent_set
-            else (1e-8, 1e-3),
-            units=units_meta("mass")
-            * units_meta("length") ** -2
-            * units_meta("time") ** -1,
-            doc="Mass flux across membrane at inlet and outlet",
-        )
-
-
-        @self.Constraint(
-            self.flowsheet().config.time,
-            self.difference_elements,
-            self.config.property_package.phase_list,
-            self.config.property_package.component_list,
-            doc="Solvent and solute mass flux",
-        )
-        def eq_flux_mass(b, t, x, p, j):
-            prop_feed = b.properties[t, x]
-            prop_perm = b.permeate_side[t, x]
-            interface = b.properties_interface[t, x]
-            comp = self.config.property_package.get_component(j)
-            if comp.is_solvent():
-                return b.flux_mass_phase_comp[t, x, p, j] == b.A_comp[
-                    t, j
-                ] * b.dens_solvent * (
-                    (prop_feed.pressure - prop_perm.pressure)
-                    - (
-                        interface.pressure_osm_phase[p]
-                        - prop_perm.pressure_osm_phase[p]
-                    )
-                )
-            elif comp.is_solute():
-                return b.flux_mass_phase_comp[t, x, p, j] == b.B_comp[t, j] * (
-                    interface.conc_mass_phase_comp[p, j]
-                    - prop_perm.conc_mass_phase_comp[p, j]
-                )
-
-        @self.Expression(
-            self.flowsheet().config.time,
-            self.config.property_package.phase_list,
-            self.config.property_package.component_list,
-            doc="Average flux expression",
-        )
-        def flux_mass_phase_comp_avg(b, t, p, j):
-            return (
-                sum(
-                    b.flux_mass_phase_comp[t, x, p, j] for x in self.difference_elements
-                )
-                / self.nfe
-            )
-
-        return self.eq_flux_mass
 
     def add_concentration_polarization(
         self,
@@ -523,36 +366,42 @@ class MembraneChannelMixin:
                     b.properties_interface[t,x].conc_mass_phase_comp["Liq", j]
                     == b.properties[t,x].conc_mass_phase_comp["Liq", j]
                 )
+            return
 
-        elif concentration_polarization_type == ConcentrationPolarizationType.fixed:
+        if concentration_polarization_type not in (ConcentrationPolarizationType.fixed, ConcentrationPolarizationType.calculated):
+            raise ConfigurationError(f"Unrecognized concentration_polarization_type {concentration_polarization_type}")
 
-            self.cp_modulus = Var(
-                self.flowsheet().config.time,
-                self.length_domain,
-                solute_set,
-                initialize=1.1,
-                bounds=(0.9, 3),
-                domain=NonNegativeReals,
-                units=pyunits.dimensionless,
-                doc="Concentration polarization modulus",
+        self.cp_modulus = Var(
+            self.flowsheet().config.time,
+            self.length_domain,
+            solute_set,
+            initialize=1.1,
+            bounds=(0.9, 3),
+            domain=NonNegativeReals,
+            units=pyunits.dimensionless,
+            doc="Concentration polarization modulus",
+        )
+
+        @self.Constraint(
+           self.flowsheet().config.time,
+           self.difference_elements,
+           solute_set,
+           doc="Concentration polarization",
+        )   
+        def eq_concentration_polarization(b, t, x, j):
+            return (
+                self.properties_interface[t,x].conc_mass_phase_comp["Liq", j]
+                == self.properties[t,x].conc_mass_phase_comp["Liq", j] * self.cp_modulus[t, x, j]
             )
 
-            @self.Constraint(
-               self.flowsheet().config.time,
-               self.difference_elements,
-               solute_set,
-               doc="Concentration polarization",
-            )   
-            def eq_concentration_polarization(b, t, x, j):
-                return (
-                    self.properties_interface[t,x].conc_mass_phase_comp["Liq", j]
-                    == self.properties[t,x].conc_mass_phase_comp["Liq", j] * self.cp_modulus[t, x, j]
-                )
-
-        elif concentration_polarization_type == ConcentrationPolarizationType.calculated:
+        if concentration_polarization_type == ConcentrationPolarizationType.calculated:
             if mass_transfer_coefficient == MassTransferCoefficient.none:
                 raise ConfigurationError()
 
+            # NOTE: calculated concentration polarization is relegated
+            #       to the unit model as it depends on flux 
+
+            # mass_transfer_coefficient is either calculated or fixed
             self.K = Var(
                 self.flowsheet().config.time,
                 self.length_domain,
@@ -561,54 +410,14 @@ class MembraneChannelMixin:
                 bounds=(1e-6, 1e-3),
                 domain=NonNegativeReals,
                 units=units_meta("length") * units_meta("time") ** -1,
-                doc="Mass transfer coefficient in membrane channel",
+                doc="Membrane channel mass transfer coefficient",
             )
-
-            @self.Constraint(
-               self.flowsheet().config.time,
-               self.difference_elements,
-               solute_set,
-               doc="Concentration polarization",
-            )   
-            def eq_concentration_polarization(b, t, x, j):
-                jw = self.flux_mass_phase_comp[t, x, "Liq", "H2O"] / self.dens_solvent
-                js = self.flux_mass_phase_comp[t, x, "Liq", j]
-                return self.properties_interface[t,x].conc_mass_phase_comp[
-                    "Liq", j
-                ] == self.properties[t,x].conc_mass_phase_comp["Liq", j] * exp(
-                    jw / self.K[t, x, j]
-                ) - js / jw * (
-                    exp(jw / self.K[t, x, j]) - 1
-                )
 
             if mass_transfer_coefficient == MassTransferCoefficient.calculated:
                 self._add_calculated_mass_transfer_coefficient()
 
-        else:
-            raise ConfigurationError()
-            
         return self.eq_concentration_polarization
 
-    def add_recovery_vol_phase(self):
-
-        self.recovery_vol_phase = Var(
-            self.flowsheet().config.time,
-            self.config.property_package.phase_list,
-            initialize=0.4,
-            bounds=(1e-2, 1 - 1e-6),
-            units=pyunits.dimensionless,
-            doc="Volumetric recovery rate",
-        )
-
-        # ==========================================================================
-        # Volumetric Recovery rate
-        @self.Constraint(self.flowsheet().config.time)
-        def eq_recovery_vol_phase(b, t):
-            return (
-                b.recovery_vol_phase[t, "Liq"]
-                == b.mixed_permeate[t].flow_vol_phase["Liq"]
-                / b.properties[t, self.first_element].flow_vol_phase["Liq"]
-            )
 
     def add_expressions(self):
         """
@@ -632,23 +441,6 @@ class MembraneChannelMixin:
             )
             def K_avg(b, t, j):
                 return sum(b.K[t, x, j] for x in self.difference_elements) / self.nfe
-
-    def _add_area_total(self):
-        units_meta = self.config.property_package.get_metadata().get_derived_units
-        self.area_total = Var(
-            initialize=10,
-            bounds=(1e-1, 1e3),
-            domain=NonNegativeReals,
-            units=units_meta("length") ** 2,
-            doc="Total Membrane Channel area",
-        )
-
-    def _add_area_total_equation(self):
-        # ==========================================================================
-        # Membrane area equation
-        @self.Constraint(doc="Total Membrane area")
-        def eq_area_total(b):
-            return b.area_total == b.length * b.width
 
     ## should be called by add concentration polarization
     def _add_calculated_mass_transfer_coefficient(self):
@@ -717,7 +509,20 @@ class MembraneChannelMixin:
         if hasattr(self, "channel_height"):
             return
 
+        if not hasattr(self, "width"):
+            raise ConfigurationError(f"Due to either a calculated mass transfer coefficient or a calculated pressure change, a ``width`` variable needs to be supplied to `add_geometry` for this MembraneChannel")
+
         units_meta = self.config.property_package.get_metadata().get_derived_units
+
+        if not hasattr(self, "area"):
+            # comes from ControlVolume1D for 1DMC
+            self.area = Var(
+                initialize=1e-3 * 1 * 0.95,
+                bounds=(0, 1e3),
+                domain=NonNegativeReals,
+                units=units_meta("length") ** 2,
+                doc="Cross sectional area",
+            )
 
         self.channel_height = Var(
             initialize=1e-3,
@@ -779,7 +584,7 @@ class MembraneChannelMixin:
                 * b.dh
             )
 
-    def _add_interface_blocks(
+    def _add_interface_stateblock(
         self, has_phase_equilibrium=None
     ):
         """
@@ -796,42 +601,12 @@ class MembraneChannelMixin:
         tmp_dict["has_phase_equilibrium"] = has_phase_equilibrium
         tmp_dict["defined_state"] = False  # these blocks are not inlets or outlets
 
-        self.permeate_side = self.config.property_package.build_state_block(
-            self.flowsheet().config.time,
-            self.length_domain,
-            doc="Material properties of permeate along permeate channel",
-            default=tmp_dict,
-        )
-        self.mixed_permeate = self.config.property_package.build_state_block(
-            self.flowsheet().config.time,
-            doc="Material properties of mixed permeate exiting the module",
-            default=tmp_dict,
-        )
-
         self.properties_interface = self.config.property_package.build_state_block(
             self.flowsheet().config.time,
             self.length_domain,
             doc="Material properties of feed-side membrane interface",
             default=tmp_dict,
         )
-
-    def _add_membrane_pressure_balances(self):
-
-        @self.Constraint(
-            self.flowsheet().config.time,
-            self.length_domain,
-            doc="Isobaric assumption for permeate out",
-        )
-        def eq_permeate_outlet_isobaric(b, t, x):
-            return b.permeate_side[t, x].pressure == b.mixed_permeate[t].pressure
-
-        @self.Constraint(
-            self.flowsheet().config.time,
-            self.difference_elements,
-            doc="Pressure at interface",
-        )
-        def eq_equal_pressure_interface(b, t, x):
-            return b.properties_interface[t, x].pressure == b.properties[t, x].pressure
 
     def _add_calculated_pressure_change(self):
         self._add_calculated_pressure_change_mass_transfer_components()
@@ -922,11 +697,6 @@ class MembraneChannelMixin:
                          property package).
         """
 
-        source = self.properties[
-            self.flowsheet().config.time.first(), self.first_element
-        ]
-        mixed_permeate_properties = self.mixed_permeate[self.flowsheet().config.time.first()]
-
         # assumptions
         if initialize_guess is None:
             initialize_guess = {}
@@ -937,13 +707,14 @@ class MembraneChannelMixin:
         if "solute_recovery" not in initialize_guess:
             initialize_guess["solute_recovery"] = 0.01
         if "cp_modulus" not in initialize_guess:
-            if (
-                self.config.concentration_polarization_type
-                != ConcentrationPolarizationType.none
-            ):
+            if hasattr(self, "cp_modulus"):
                 initialize_guess["cp_modulus"] = 1.1
             else:
                 initialize_guess["cp_modulus"] = 1
+
+        source = self.properties[
+            self.flowsheet().config.time.first(), self.first_element
+        ]
 
         if state_args is None:
             state_args = {}
@@ -968,24 +739,16 @@ class MembraneChannelMixin:
 
         # slightly modify initial values for other state blocks
         state_args_retentate = deepcopy(state_args)
-        state_args_permeate = deepcopy(state_args)
 
         state_args_retentate["pressure"] += initialize_guess["deltaP"]
-        state_args_permeate["pressure"] = mixed_permeate_properties.pressure.value
         for j in self.config.property_package.solvent_set:
             state_args_retentate["flow_mass_phase_comp"][("Liq", j)] *= (
                 1 - initialize_guess["solvent_recovery"]
             )
-            state_args_permeate["flow_mass_phase_comp"][("Liq", j)] *= initialize_guess[
-                "solvent_recovery"
-            ]
         for j in self.config.property_package.solute_set:
             state_args_retentate["flow_mass_phase_comp"][("Liq", j)] *= (
                 1 - initialize_guess["solute_recovery"]
             )
-            state_args_permeate["flow_mass_phase_comp"][("Liq", j)] *= initialize_guess[
-                "solute_recovery"
-            ]
 
         state_args_interface_in = deepcopy(state_args)
         state_args_interface_out = deepcopy(state_args_retentate)
@@ -999,30 +762,30 @@ class MembraneChannelMixin:
             ] *= initialize_guess["cp_modulus"]
 
         # TODO: I think this is what we'd like to do, but IDAES needs to be changed
-        state_args_interface = {}
-        for t in self.flowsheet().config.time:
-            for x in self.length_domain:
-                assert 0.0 <= x <= 1.0
-                state_args_tx = {}
-                for k in state_args_interface_in:
-                    if isinstance(state_args_interface_in[k], dict):
-                        if k not in state_args_tx:
-                            state_args_tx[k] = {}
-                        for index in state_args_interface_in[k]:
-                            state_args_tx[k][index] = (
-                                1.0 - x
-                            ) * state_args_interface_in[k][
-                                index
-                            ] + x * state_args_interface_out[
-                                k
-                            ][
-                                index
-                            ]
-                    else:
-                        state_args_tx[k] = (1.0 - x) * state_args_interface_in[
-                            k
-                        ] + x * state_args_interface_out[k]
-                state_args_interface[t, x] = state_args_tx
+        # state_args_interface = {}
+        # for t in self.flowsheet().config.time:
+        #     for x in self.length_domain:
+        #         assert 0.0 <= x <= 1.0
+        #         state_args_tx = {}
+        #         for k in state_args_interface_in:
+        #             if isinstance(state_args_interface_in[k], dict):
+        #                 if k not in state_args_tx:
+        #                     state_args_tx[k] = {}
+        #                 for index in state_args_interface_in[k]:
+        #                     state_args_tx[k][index] = (
+        #                         1.0 - x
+        #                     ) * state_args_interface_in[k][
+        #                         index
+        #                     ] + x * state_args_interface_out[
+        #                         k
+        #                     ][
+        #                         index
+        #                     ]
+        #             else:
+        #                 state_args_tx[k] = (1.0 - x) * state_args_interface_in[
+        #                     k
+        #                 ] + x * state_args_interface_out[k]
+        #         state_args_interface[t, x] = state_args_tx
 
         x = 0.5
         state_args_tx = {}
@@ -1040,11 +803,27 @@ class MembraneChannelMixin:
                 ] + x * state_args_interface_out[k]
         state_args_interface = state_args_tx
 
+        x = 0.5
+        state_args_cv = {}
+        for k in state_args:
+            if isinstance(state_args[k], dict):
+                if k not in state_args_cv:
+                    state_args_cv[k] = {}
+                for index in state_args[k]:
+                    state_args_cv[k][index] = (1.0 - x) * state_args[k][
+                        index
+                    ] + x * state_args_retentate[k][index]
+            else:
+                state_args_cv[k] = (1.0 - x) * state_args[
+                    k
+                ] + x * state_args_retentate[k]
+        state_args_interface = state_args_tx
+
         return {
             "feed_side": state_args,
             "retentate": state_args_retentate,
-            "permeate": state_args_permeate,
             "interface": state_args_interface,
+            "control_volume": state_args_cv,
         }
 
     def initialize(
@@ -1097,13 +876,13 @@ class MembraneChannelMixin:
         opt = get_solver(solver, optarg)
 
         # Get inlet state if not provided
-        init_log = idaeslog.getInitLogger(self.name, outlvl, tag="membrane_channel")
-        solve_log = idaeslog.getSolveLogger(self.name, outlvl, tag="membrane_channel")
+        init_log = idaeslog.getInitLogger(self.name, outlvl, tag="control_volume")
+        solve_log = idaeslog.getSolveLogger(self.name, outlvl, tag="control_volume")
 
         state_args = self._get_state_args(initialize_guess, state_args)
 
         # intialize self.properties
-        source_flags = super().initialize(state_args=state_args["feed_side"], outlvl=outlvl, optarg=optarg, solver=solver, hold_state=True)
+        source_flags = super().initialize(state_args=state_args["control_volume"], outlvl=outlvl, optarg=optarg, solver=solver, hold_state=True)
 
         self.properties_interface.initialize(
             outlvl=outlvl,
@@ -1112,26 +891,12 @@ class MembraneChannelMixin:
             state_args=state_args["interface"],
         )
 
-        self.permeate_side.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["permeate"],
-        )
-
-        self.mixed_permeate.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["permeate"],
-        )
-
         with idaeslog.solver_log(solve_log, idaeslog.DEBUG) as slc:
             res = opt.solve(self, tee=slc.tee)
             # occasionally it might be worth retrying a solve
             if not check_optimal_termination(res):
                 init_log.warning(
-                    "Trouble solving ReverseOsmosis1D unit model, trying one more time"
+                    f"Trouble solving membrane channel {self.name}, trying one more time"
                 )
                 res = opt.solve(self, tee=slc.tee)
         if not check_optimal_termination(res):
@@ -1139,136 +904,13 @@ class MembraneChannelMixin:
 
         init_log.info("Initialization Complete")
 
-        self.release_state(source_flags, outlvl)
-    
-
-
-    def _get_performance_contents(self, time_point=0):
-        x_in = self.first_element
-        x_interface_in = self.difference_elements.first()
-        x_out = self.length_domain.last()
-        feed_inlet = self.properties[time_point, x_in]
-        feed_outlet = self.properties[time_point, x_out]
-        interface_inlet = self.properties_interface[
-            time_point, x_interface_in
-        ]
-        interface_outlet = self.properties_interface[time_point, x_out]
-        permeate = self.mixed_permeate[time_point]
-        var_dict = {}
-        expr_dict = {}
-        var_dict["Volumetric Recovery Rate"] = self.recovery_vol_phase[
-            time_point, "Liq"
-        ]
-        var_dict["Solvent Mass Recovery Rate"] = self.recovery_mass_phase_comp[
-            time_point, "Liq", "H2O"
-        ]
-        var_dict["Membrane Area"] = self.area_total
-        if hasattr(self, "length"):
-            var_dict["Membrane Length"] = self.length
-        if hasattr(self, "width"):
-            var_dict["Membrane Width"] = self.width
-        if hasattr(self, "pressure_change_total"):
-            var_dict["Pressure Change"] = self.pressure_change_total[time_point]
-        if hasattr(self, "N_Re"):
-            var_dict["Reynolds Number @Inlet"] = self.N_Re[time_point, x_in]
-            var_dict["Reynolds Number @Outlet"] = self.N_Re[time_point, x_out]
-        if hasattr(self, "velocity"):
-            var_dict["Velocity @Inlet"] = self.velocity[time_point, x_in]
-            var_dict["Velocity @Outlet"] = self.velocity[time_point, x_out]
-        for j in self.config.property_package.solute_set:
-            if interface_inlet.is_property_constructed("conc_mass_phase_comp"):
-                var_dict[
-                    f"{j} Concentration @Inlet,Membrane-Interface "
-                ] = interface_inlet.conc_mass_phase_comp["Liq", j]
-            if interface_outlet.is_property_constructed("conc_mass_phase_comp"):
-                var_dict[
-                    f"{j} Concentration @Outlet,Membrane-Interface "
-                ] = interface_outlet.conc_mass_phase_comp["Liq", j]
-            if feed_inlet.is_property_constructed("conc_mass_phase_comp"):
-                var_dict[
-                    f"{j} Concentration @Inlet,Bulk"
-                ] = feed_inlet.conc_mass_phase_comp["Liq", j]
-            if feed_outlet.is_property_constructed("conc_mass_phase_comp"):
-                var_dict[
-                    f"{j} Concentration @Outlet,Bulk"
-                ] = feed_outlet.conc_mass_phase_comp["Liq", j]
-            if permeate.is_property_constructed("conc_mass_phase_comp"):
-                var_dict[f"{j} Permeate Concentration"] = permeate.conc_mass_phase_comp[
-                    "Liq", j
-                ]
-        if interface_outlet.is_property_constructed("pressure_osm_phase"):
-            var_dict[
-                "Osmotic Pressure @Outlet,Membrane-Interface "
-            ] = interface_outlet.pressure_osm_phase["Liq"]
-        if feed_outlet.is_property_constructed("pressure_osm_phase"):
-            var_dict["Osmotic Pressure @Outlet,Bulk"] = feed_outlet.pressure_osm_phase[
-                "Liq"
-            ]
-        if interface_inlet.is_property_constructed("pressure_osm_phase"):
-            var_dict[
-                "Osmotic Pressure @Inlet,Membrane-Interface"
-            ] = interface_inlet.pressure_osm_phase["Liq"]
-        if feed_inlet.is_property_constructed("pressure_osm_phase"):
-            var_dict["Osmotic Pressure @Inlet,Bulk"] = feed_inlet.pressure_osm_phase[
-                "Liq"
-            ]
-        if feed_inlet.is_property_constructed("flow_vol_phase"):
-            var_dict["Volumetric Flowrate @Inlet"] = feed_inlet.flow_vol_phase["Liq"]
-        if feed_outlet.is_property_constructed("flow_vol_phase"):
-            var_dict["Volumetric Flowrate @Outlet"] = feed_outlet.flow_vol_phase["Liq"]
-        if hasattr(self, "dh"):
-            var_dict["Hydraulic Diameter"] = self.dh
-
-        expr_dict["Average Solvent Flux (LMH)"] = (
-            self.flux_mass_phase_comp_avg[time_point, "Liq", "H2O"] * 3.6e3
-        )
-        expr_dict["Average Reynolds Number"] = self.N_Re_avg[time_point]
-        for j in self.config.property_package.solute_set:
-            expr_dict[f"{j} Average Solute Flux (GMH)"] = (
-                self.flux_mass_phase_comp_avg[time_point, "Liq", j] * 3.6e6
-            )
-            expr_dict[f"{j} Average Mass Transfer Coefficient (mm/h)"] = (
-                self.K_avg[time_point, j] * 3.6e6
-            )
-
-        # TODO: add more vars
-        return {"vars": var_dict, "exprs": expr_dict}
-
-    # permeate properties need to rescale solute values by 100
-    def _rescale_permeate_variable(self, var, factor=100):
-        if var not in self._permeate_scaled_properties:
-            sf = iscale.get_scaling_factor(var)
-            iscale.set_scaling_factor(var, sf * factor)
-            self._permeate_scaled_properties.add(var)
+        if hold_state:
+            return source_flags 
+        else:
+            self.release_state(source_flags, outlvl)
 
     def calculate_scaling_factors(self):
         super().calculate_scaling_factors()
-
-        # these variables should have user input, if not there will be a warning
-        if iscale.get_scaling_factor(self.area_total) is None:
-            sf = iscale.get_scaling_factor(self.area_total, default=10, warning=True)
-            iscale.set_scaling_factor(self.area_total, sf)
-
-        if iscale.get_scaling_factor(self.A_comp) is None:
-            iscale.set_scaling_factor(self.A_comp, 1e12)
-
-        if iscale.get_scaling_factor(self.B_comp) is None:
-            iscale.set_scaling_factor(self.B_comp, 1e8)
-
-        if iscale.get_scaling_factor(self.recovery_vol_phase) is None:
-            iscale.set_scaling_factor(self.recovery_vol_phase, 1)
-
-        for (t, p, j), v in self.recovery_mass_phase_comp.items():
-            if j in self.config.property_package.solvent_set:
-                sf = 1
-            elif j in self.config.property_package.solute_set:
-                sf = 100
-            if iscale.get_scaling_factor(v) is None:
-                iscale.set_scaling_factor(v, sf)
-
-        for v in self.rejection_phase_comp.values():
-            if iscale.get_scaling_factor(v) is None:
-                iscale.set_scaling_factor(v, 1)
 
         if hasattr(self, "channel_height"):
             if iscale.get_scaling_factor(self.channel_height) is None:
@@ -1316,50 +958,6 @@ class MembraneChannelMixin:
             for v in self.cp_modulus.values():
                 if iscale.get_scaling_factor(v) is None:
                     iscale.set_scaling_factor(v, 1)
-
-        # for self._rescale_permeate_variable
-        self._permeate_scaled_properties = ComponentSet()
-
-        for sb in (self.permeate_side, self.mixed_permeate):
-            for blk in sb.values():
-                for j in self.config.property_package.solute_set:
-                    self._rescale_permeate_variable(blk.flow_mass_phase_comp["Liq", j])
-                    if blk.is_property_constructed("mass_frac_phase_comp"):
-                        self._rescale_permeate_variable(
-                            blk.mass_frac_phase_comp["Liq", j]
-                        )
-                    if blk.is_property_constructed("conc_mass_phase_comp"):
-                        self._rescale_permeate_variable(
-                            blk.conc_mass_phase_comp["Liq", j]
-                        )
-                    if blk.is_property_constructed("mole_frac_phase_comp"):
-                        self._rescale_permeate_variable(blk.mole_frac_phase_comp[j])
-                    if blk.is_property_constructed("molality_phase_comp"):
-                        self._rescale_permeate_variable(
-                            blk.molality_phase_comp["Liq", j]
-                        )
-                if blk.is_property_constructed("pressure_osm_phase"):
-                    self._rescale_permeate_variable(blk.pressure_osm_phase["Liq"])
-
-        for (t, x, p, j), v in self.flux_mass_phase_comp.items():
-            if iscale.get_scaling_factor(v) is None:
-                comp = self.config.property_package.get_component(j)
-                if comp.is_solvent():  # scaling based on solvent flux equation
-                    sf = (
-                        iscale.get_scaling_factor(self.A_comp[t, j])
-                        * iscale.get_scaling_factor(self.dens_solvent)
-                        * iscale.get_scaling_factor(
-                            self.properties[t, x].pressure
-                        )
-                    )
-                    iscale.set_scaling_factor(v, sf)
-                elif comp.is_solute():  # scaling based on solute flux equation
-                    sf = iscale.get_scaling_factor(
-                        self.B_comp[t, j]
-                    ) * iscale.get_scaling_factor(
-                        self.properties[t, x].conc_mass_phase_comp[p, j]
-                    )
-                    iscale.set_scaling_factor(v, sf)
 
 
 # helper for validating configuration arguments for this CV
