+++ watertap/unit_models/reverse_osmosis_1D.py
@@ -27,12 +27,13 @@ from idaes.core import (
     declare_process_block_class,
     useDefault,
 )
+from idaes.core.util import scaling as iscale
 from idaes.core.util.misc import add_object_reference
 import idaes.logger as idaeslog
 
-from watertap.core import MembraneChannel1D
+from watertap.core import MembraneChannel1DBlock, ConcentrationPolarizationType, MassTransferCoefficient, PressureChangeType
 from watertap.core.membrane_channel1d import CONFIG_Template
-from watertap.unit_models.reverse_osmosis_base import ReverseOsmosisBaseData
+from watertap.unit_models.reverse_osmosis_base import ReverseOsmosisBaseData, _add_has_full_reporting
 
 __author__ = "Adam Atia"
 
@@ -45,6 +46,8 @@ class ReverseOsmosis1DData(ReverseOsmosisBaseData):
     """Standard 1D Reverse Osmosis Unit Model Class."""
 
     CONFIG = CONFIG_Template()
+
+    _add_has_full_reporting(CONFIG)
 
     def _process_config(self):
         if self.config.transformation_method is useDefault:
@@ -67,9 +70,12 @@ class ReverseOsmosis1DData(ReverseOsmosisBaseData):
             )
             self.config.transformation_scheme = "BACKWARD"
 
-    def _add_membrane_channel(self):
+    def _add_feed_side_membrane_channel(self):
+        # Check configuration errors
+        self._process_config()
+
         # Build 1D Membrane Channel
-        self.membrane_channel = MembraneChannel1D(
+        self.feed_side = MembraneChannel1DBlock(
             default={
                 "dynamic": self.config.dynamic,
                 "has_holdup": self.config.has_holdup,
@@ -83,6 +89,12 @@ class ReverseOsmosis1DData(ReverseOsmosisBaseData):
             }
         )
 
+        self._add_area()
+        self._add_length_and_width()
+
+        self.feed_side.add_geometry(length_var=self.length, width_var=self.width)
+
+
     def build(self):
         """
         Build 1D RO model (pre-DAE transformation).
@@ -93,11 +105,121 @@ class ReverseOsmosis1DData(ReverseOsmosisBaseData):
         Returns:
             None
         """
-        # Check configuration errors
-        self._process_config()
-
         # Call UnitModel.build to setup dynamics
         super().build()
 
+
+    def _add_mass_transfer(self):
+
+        units_meta = self.config.property_package.get_metadata().get_derived_units
+
+        def mass_transfer_phase_comp_initialize(b, t, x, p, j):
+            return value(
+                self.feed_side.properties[t, x].get_material_flow_terms("Liq", j)
+                * self.recovery_mass_phase_comp[t, "Liq", j]
+            )
+
+        self.mass_transfer_phase_comp = Var(
+            self.flowsheet().config.time,
+            self.length_domain,
+            self.config.property_package.phase_list,
+            self.config.property_package.component_list,
+            initialize=mass_transfer_phase_comp_initialize,
+            bounds=(0.0, 1e6),
+            domain=NonNegativeReals,
+            units=units_meta("mass")
+            * units_meta("time") ** -1
+            * units_meta("length") ** -1,
+            doc="Mass transfer to permeate",
+        )
+
         # ==========================================================================
-        """ Add references to control volume geometry."""
+        # Mass transfer term equation
+        @self.Constraint(
+            self.flowsheet().config.time,
+            self.difference_elements,
+            self.config.property_package.phase_list,
+            self.config.property_package.component_list,
+            doc="Mass transfer term",
+        )
+        def eq_mass_transfer_term(b, t, x, p, j):
+            return (
+                b.mass_transfer_phase_comp[t, x, p, j]
+                == -b.feed_side.mass_transfer_term[t, x, p, j]
+            )
+
+        # ==========================================================================
+        # Feed and permeate-side mass transfer connection --> Mp,j = Mf,transfer = Jj * W * L/n
+        @self.Constraint(
+            self.flowsheet().config.time,
+            self.difference_elements,
+            self.config.property_package.phase_list,
+            self.config.property_package.component_list,
+            doc="Mass transfer from feed to permeate",
+        )
+        def eq_connect_mass_transfer(b, t, x, p, j):
+            return (
+                b.permeate_side[t, x].get_material_flow_terms(p, j)
+                == -b.feed_side.mass_transfer_term[t, x, p, j] * b.length / b.nfe
+            )
+
+        # ==========================================================================
+        # Mass flux = feed mass transfer equation
+        @self.Constraint(
+            self.flowsheet().config.time,
+            self.difference_elements,
+            self.config.property_package.phase_list,
+            self.config.property_package.component_list,
+            doc="Mass transfer term",
+        )
+        def eq_mass_flux_equal_mass_transfer(b, t, x, p, j):
+            return (
+                b.flux_mass_phase_comp[t, x, p, j] * b.width
+                == -b.feed_side.mass_transfer_term[t, x, p, j]
+            )
+
+        # ==========================================================================
+        # Final permeate mass flow rate (of solvent and solute) --> Mp,j, final = sum(Mp,j)
+
+        @self.Constraint(
+            self.flowsheet().config.time,
+            self.config.property_package.phase_list,
+            self.config.property_package.component_list,
+            doc="Permeate mass flow rates exiting unit",
+        )
+        def eq_permeate_production(b, t, p, j):
+            return b.mixed_permeate[t].get_material_flow_terms(p, j) == sum(
+                b.permeate_side[t, x].get_material_flow_terms(p, j)
+                for x in b.difference_elements
+            )
+
+    def calculate_scaling_factors(self):
+        if iscale.get_scaling_factor(self.dens_solvent) is None:
+            sf = iscale.get_scaling_factor(
+                self.feed_side.properties[0, 0].dens_mass_phase["Liq"]
+            )
+            iscale.set_scaling_factor(self.dens_solvent, sf)
+
+        super().calculate_scaling_factors()
+
+        # these variables should have user input, if not there will be a warning
+        if iscale.get_scaling_factor(self.width) is None:
+            sf = iscale.get_scaling_factor(self.width, default=1, warning=True)
+            iscale.set_scaling_factor(self.width, sf)
+
+        if iscale.get_scaling_factor(self.length) is None:
+            sf = iscale.get_scaling_factor(self.length, default=10, warning=True)
+            iscale.set_scaling_factor(self.length, sf)
+
+        for (t, x, p, j), v in self.mass_transfer_phase_comp.items():
+            sf = (
+                iscale.get_scaling_factor(
+                    self.feed_side.properties[t, x].get_material_flow_terms(p, j)
+                )
+                / iscale.get_scaling_factor(self.feed_side.length)
+            ) * value(self.nfe)
+            if iscale.get_scaling_factor(v) is None:
+                iscale.set_scaling_factor(v, sf)
+            v = self.feed_side.mass_transfer_term[t, x, p, j]
+            if iscale.get_scaling_factor(v) is None:
+                iscale.set_scaling_factor(v, sf)
