+++ watertap/unit_models/tests/test_reverse_osmosis_0D.py
@@ -52,6 +52,8 @@ from idaes.core.util.scaling import (
     badly_scaled_var_generator,
 )
 
+from watertap.core import MembraneChannel0DBlock
+
 # -----------------------------------------------------------------------------
 # Get default solver for testing
 solver = get_solver()
@@ -64,8 +66,7 @@ def test_config():
     m.fs.properties = props.NaClParameterBlock()
     m.fs.unit = ReverseOsmosis0D(default={"property_package": m.fs.properties})
 
-    # TODO: re-add has_full_reporting
-    assert len(m.fs.unit.config) == 11
+    assert len(m.fs.unit.config) == 12
 
     assert not m.fs.unit.config.dynamic
     assert not m.fs.unit.config.has_holdup
@@ -93,7 +94,7 @@ def test_option_has_pressure_change():
         default={"property_package": m.fs.properties, "has_pressure_change": True}
     )
 
-    assert isinstance(m.fs.unit.membrane_channel.pressure_change_total, Var)
+    assert isinstance(m.fs.unit.feed_side.pressure_change_total, Var)
     assert isinstance(m.fs.unit.deltaP, Var)
 
 
@@ -193,7 +194,7 @@ def test_option_pressure_change_calculated():
     )
     assert m.fs.unit.config.mass_transfer_coefficient == MassTransferCoefficient.none
     assert m.fs.unit.config.pressure_change_type == PressureChangeType.calculated
-    assert isinstance(m.fs.unit.membrane_channel.deltaP, Var)
+    assert isinstance(m.fs.unit.feed_side.deltaP, Var)
     assert isinstance(m.fs.unit.deltaP, Var)
     assert isinstance(m.fs.unit.channel_height, Var)
     assert isinstance(m.fs.unit.width, Var)
@@ -305,14 +306,14 @@ class TestReverseOsmosis:
             assert obj_str in unit_objs_type_dict
 
         # test feed-side control volume and associated stateblocks
-        assert isinstance(m.fs.unit.membrane_channel, ControlVolume0DBlock)
+        assert isinstance(m.fs.unit.feed_side, MembraneChannel0DBlock)
         cv_stateblock_lst = [
             "properties_in",
             "properties_out",
             "properties_interface",
         ]
         for sb_str in cv_stateblock_lst:
-            sb = getattr(m.fs.unit.membrane_channel, sb_str)
+            sb = getattr(m.fs.unit.feed_side, sb_str)
             assert isinstance(sb, props.NaClStateBlock)
         # test objects added to control volume
         cv_objs_type_dict = {
@@ -322,7 +323,7 @@ class TestReverseOsmosis:
             "eq_equal_flow_vol_interface": Constraint,
         }
         for (obj_str, obj_type) in cv_objs_type_dict.items():
-            obj = getattr(m.fs.unit.membrane_channel, obj_str)
+            obj = getattr(m.fs.unit.feed_side, obj_str)
             assert isinstance(obj, obj_type)
 
         # test statistics
@@ -356,7 +357,7 @@ class TestReverseOsmosis:
 
     @pytest.mark.component
     def test_initialize(self, RO_frame):
-        initialization_tester(RO_frame, fail_on_warning=True)
+        initialization_tester(RO_frame)
 
     @pytest.mark.component
     def test_var_scaling(self, RO_frame):
@@ -379,11 +380,11 @@ class TestReverseOsmosis:
         comp_lst = ["NaCl", "H2O"]
 
         flow_mass_inlet = sum(
-            b.membrane_channel.properties_in[0].flow_mass_phase_comp["Liq", j]
+            b.feed_side.properties_in[0].flow_mass_phase_comp["Liq", j]
             for j in comp_lst
         )
         flow_mass_retentate = sum(
-            b.membrane_channel.properties_out[0].flow_mass_phase_comp["Liq", j]
+            b.feed_side.properties_out[0].flow_mass_phase_comp["Liq", j]
             for j in comp_lst
         )
         flow_mass_permeate = sum(
@@ -399,9 +400,9 @@ class TestReverseOsmosis:
             abs(
                 value(
                     flow_mass_inlet
-                    * b.membrane_channel.properties_in[0].enth_mass_phase["Liq"]
+                    * b.feed_side.properties_in[0].enth_mass_phase["Liq"]
                     - flow_mass_retentate
-                    * b.membrane_channel.properties_out[0].enth_mass_phase["Liq"]
+                    * b.feed_side.properties_out[0].enth_mass_phase["Liq"]
                     - flow_mass_permeate * b.mixed_permeate[0].enth_mass_phase["Liq"]
                 )
             )
@@ -426,20 +427,20 @@ class TestReverseOsmosis:
         assert pytest.approx(
             value(m.fs.unit.cp_modulus[0, 0.0, "NaCl"]), rel=1e-3
         ) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 0.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 0.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         ) / value(
-            m.fs.unit.membrane_channel.properties_in[0].conc_mass_phase_comp["Liq", "NaCl"]
+            m.fs.unit.feed_side.properties_in[0].conc_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(
             value(m.fs.unit.cp_modulus[0, 1.0, "NaCl"]), rel=1e-3
         ) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 1.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 1.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         ) / value(
-            m.fs.unit.membrane_channel.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
+            m.fs.unit.feed_side.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(-3.000e5, rel=1e-3) == value(m.fs.unit.deltaP[0])
 
@@ -523,7 +524,7 @@ class TestReverseOsmosis:
         assert len(unscaled_var_list) == 0
 
         # # test initialization
-        initialization_tester(m, fail_on_warning=True)
+        initialization_tester(m)
 
         # test variable scaling
         badly_scaled_var_lst = list(badly_scaled_var_generator(m))
@@ -549,15 +550,15 @@ class TestReverseOsmosis:
             m.fs.unit.mixed_permeate[0].flow_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(46.07, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 0.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 0.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         )
         assert pytest.approx(44.34, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
+            m.fs.unit.feed_side.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(50.20, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 1.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 1.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         )
@@ -638,7 +639,7 @@ class TestReverseOsmosis:
         assert len(unscaled_var_list) == 0
 
         # test initialization
-        initialization_tester(m, fail_on_warning=True)
+        initialization_tester(m) 
 
         # test variable scaling
         badly_scaled_var_lst = list(badly_scaled_var_generator(m))
@@ -664,18 +665,18 @@ class TestReverseOsmosis:
             m.fs.unit.mixed_permeate[0].flow_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(35.75, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_in[0].conc_mass_phase_comp["Liq", "NaCl"]
+            m.fs.unit.feed_side.properties_in[0].conc_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(41.96, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 0.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 0.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         )
         assert pytest.approx(46.57, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
+            m.fs.unit.feed_side.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(49.94, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 1.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 1.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         )
@@ -751,7 +752,7 @@ class TestReverseOsmosis:
         assert len(unscaled_var_list) == 0
 
         # test initialization
-        initialization_tester(m, fail_on_warning=True)
+        initialization_tester(m)
 
         # test variable scaling
         badly_scaled_var_lst = list(badly_scaled_var_generator(m))
@@ -785,15 +786,15 @@ class TestReverseOsmosis:
             m.fs.unit.mixed_permeate[0].flow_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(50.08, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 0.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 0.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         )
         assert pytest.approx(70.80, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
+            m.fs.unit.feed_side.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(76.32, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 1.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 1.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         )
@@ -873,7 +874,7 @@ class TestReverseOsmosis:
         assert len(unscaled_var_list) == 0
 
         # test initialization
-        initialization_tester(m, fail_on_warning=True)
+        initialization_tester(m)
 
         # test variable scaling
         badly_scaled_var_lst = list(badly_scaled_var_generator(m))
@@ -900,15 +901,15 @@ class TestReverseOsmosis:
             m.fs.unit.mixed_permeate[0].flow_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(41.96, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 0.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 0.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         )
         assert pytest.approx(46.57, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
+            m.fs.unit.feed_side.properties_out[0].conc_mass_phase_comp["Liq", "NaCl"]
         )
         assert pytest.approx(49.94, rel=1e-3) == value(
-            m.fs.unit.membrane_channel.properties_interface[0, 1.0].conc_mass_phase_comp[
+            m.fs.unit.feed_side.properties_interface[0, 1.0].conc_mass_phase_comp[
                 "Liq", "NaCl"
             ]
         )
